+++ app/Http/Controllers/Controller.php
@@ -10,4 +10,25 @@ use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 class Controller extends BaseController
 {
     use AuthorizesRequests, DispatchesJobs, ValidatesRequests;
+
+    /**
+     * @var string
+     */
+    private $gateway;
+
+    /**
+     * @param $gateway
+     */
+    protected function setGateway($gateway)
+    {
+        $this->gateway = $gateway;
+    }
+
+    /**
+     * @return string
+     */
+    protected function getGateway()
+    {
+        return $this->gateway;
+    }
 }