+++ resources/lang/en/common.php
@@ -1,0 +1,16 @@
+<?php
+
+return [
+    'change' => 'change',
+    'confirm' => 'confirm',
+    'ip_address' => 'ip address',
+    'last_update' => 'last update',
+    'new' => 'new',
+    'new_f' => 'new',
+    'old' => 'old',
+    'old_f' => 'old',
+    'old_password_doesnt_match' => 'old password is wrong',
+    'operation_completed_successfully' => 'operation completed successfully',
+    'settings' => 'settings',
+    'update' => 'update',
+];