+++ resources/lang/it/common.php
@@ -1,0 +1,16 @@
+<?php
+
+return [
+    'change' => 'cambio',
+    'confirm' => 'conferma',
+    'ip_address' => 'indirizzo IP',
+    'last_update' => 'ultimo aggiornamento',
+    'new' => 'nuovo',
+    'new_f' => 'nuova',
+    'old' => 'vecchio',
+    'old_f' => 'vecchia',
+    'old_password_doesnt_match' => 'la vecchia password è errata',
+    'operation_completed_successfully' => 'operazione completata con successo',
+    'settings' => 'impostazioni',
+    'update' => 'aggiorna',
+];