+++ resources/lang/it/validation.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'confirmed' => 'conferma :attribute non corrisponde al precedente',
+];