+++ routes/web.php
@@ -11,6 +11,11 @@
 |
 */
 
-Route::get('/', function () {
-    return view('welcome');
-});
+Route::get('/', 'Auth\LoginController@showLoginForm')->name('login');
+Route::post('/', 'Auth\LoginController@login')->name('login');
+Route::post('/logout', 'Auth\LoginController@logout')->name('logout');
+
+Route::get('/dashboard', 'HomeController@dashboard')->name('dashboard');
+
+Route::get('/user/password', 'Settings\PasswordController@show')->name('password_show');
+Route::patch('/user/password', 'Settings\PasswordController@update')->name('password_update');