+++ webpack.mix.js
@@ -11,5 +11,10 @@ const mix = require('laravel-mix');
  |
  */
 
-mix.js('resources/js/app.js', 'public/js')
-   .sass('resources/sass/app.scss', 'public/css');
+mix.js([
+        'resources/js/app.js',
+        'resources/js/logout.js',
+        'resources/js/theme.v200.js'
+    ],'public/js')
+    .styles('resources/css/theme.v200.css', 'public/css/theme.css')
+    .sass('resources/sass/app.scss', 'public/css')