+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.0
+# 29-03-2019
++ Add bootstrap layout
++ Add Public IP on dashboard
++ Enable change password
+
 # 0.1.0
 ## 24-03-2019
 + First boot
++ Add get public IP